Utility function for cross filter data mask
